+++ extension/toolbox/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to the "toolbox" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.1.0] - 2023-9-26 (+0800)
+- ⭐ add base64 encode/decode
+- 🐛 fix retain context when hidden tabs
+
 ## [1.0.2] - 2023-9-8 (+0800)
 - 🗒️docs: fix link